RealmeParts: Create HBM Service
Well in realme X2 HBM toggle without the need of service works as it should but in X2 Pro HBM acts weirdly. HBM gets disabled if device is unlocked, probably FOD is messing with HBM.
So in order to fix it we are using ACTION_USER_PRESENT intent, HBM node gets written when device is unlocked & HBM gets disabled if screen goes off.

Signed-off-by: 4j17h <7hou74m4j17h@gmail.com>
Change-Id: Ib6d6544c2a0a49a9f242e1dcb2715c9ad9ab0c27